# use eval to improve performance of dist op