### PR types
<!-- One of [ New features | Bug fixes | Function optimization | Performance optimization | Breaking changes | Others ] --> Performance optimization

### PR changes
<!-- One of [ OPs | APIs | Docs | Others ] -->  OPs

### Describe
<!-- Describe what this PR does --> use eval to improve performance
 **CPU Performance**
|op|input shape|p|before|after| speed up |
|---|---|---|---|---|---|
|dist_grad|[1, 64, 128, 128]|2| 1037.24 ms|493.03 ms|2倍| 
|dist_grad|[1, 64, 128, 128]|-inf|1002.08 ms|490.838 ms|2倍|